# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Execute Query API
  version: 1.0.0
  description: Endpoints to run saved queries and inspect results.

servers:
  - url: http://localhost:8080/api
    description: Local development server

paths:
  /execute-query/files:
    get:
      operationId: listFiles
      summary: listing all files
      tags: [ExecuteQuery]
      responses:
        '200':
          description: Successful execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
              examples:
                sample:
                  summary: Query execution result
                  value:
                    success: true
                    data: [ { "get-user-data"} ]
                    execution_time_ms: 120
        '404':
            description: No any files existing
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorDto'

  /execute-query:
    get:
      operationId: executeQuery
      summary: executing the query
      tags: [ExecuteQuery]
      parameters:
        - name: query_identifier
          in: query
          required: true
          schema:
            type: string
            example: "get_user_data"
      responses:
        '200':
          description: Successful execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
              examples:
                sample:
                  summary: Query execution result
                  value:
                    success: true
                    data: [{"id": 1, "name": "John"}, {"id": 2, "name": "Jane"}]
                    execution_time_ms: 120
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'


components:
  schemas:
    QueryResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the query execution was successful
        data:
          type: array
          items:
            type: object
          description: Array of query result rows (only present when success=true)
        error:
          type: string
          example: "Query file not found"
          description: Error message (only present when success=false)
        code:
          type: integer
          example: 404
          description: Error code (only present when success=false)
        execution_time_ms:
          type: integer
          format: int64
          description: Query execution time in milliseconds
      required:
        - success
        - execution_time_ms
    ErrorDto:
      type: object
      properties:
        errorCode:
          type: integer
        message:
          type: string